refactor: YKF関連のファイルを整理し、コントローラーに移行。不要なファイルを削除。
