+++ src/ykf/controllers/detail-controller.js
